apps/intake/backend/requirements.txt
# FastAPI and server
fastapi>=0.109.2,<1.0
uvicorn[standard]>=0.27.1,<1.0
python-multipart>=0.0.9,<1.0

# MongoDB
motor>=3.3.2,<4.0
pymongo>=4.6.1,<5.0

# Validation and settings
pydantic>=2.6.1,<3.0
pydantic-settings>=2.1.0,<3.0
email-validator>=2.1.0,<3.0

# HTTP client
httpx>=0.26.0,<1.0

# Authentication and security
python-jose[cryptography]>=3.3.0,<4.0
passlib[bcrypt]>=1.7.4,<2.0
bcrypt>=4.1.2,<5.0

# Date utilities
python-dateutil>=2.8.2,<3.0

# Testing
pytest>=7.4.4,<9.0
pytest-asyncio>=0.23.4,<1.0
pytest-cov>=4.1.0,<6.0
